+++ packages/core/src/request.ts
@@ -1,11 +1,42 @@
-import { Event, SentryRequest, Session } from '@sentry/types';
+import { Event, SdkInfo, SentryRequest, Session } from '@sentry/types';
 
 import { API } from './api';
 
+/** Extract sdk info from from the API metadata */
+function getSdkMetadataForEnvelopeHeader(api: API): SdkInfo | undefined {
+  if (!api.metadata || !api.metadata.sdk) {
+    return;
+  }
+  const { name, version } = api.metadata.sdk;
+  return { name, version };
+}
+
+/**
+ * Apply SdkInfo (name, version, packages, integrations) to the corresponding event key.
+ * Merge with existing data if any.
+ **/
+function enhanceEventWithSdkInfo(event: Event, sdkInfo?: SdkInfo): Event {
+  if (!sdkInfo) {
+    return event;
+  }
+
+  event.sdk = event.sdk || {
+    name: sdkInfo.name,
+    version: sdkInfo.version,
+  };
+  event.sdk.name = event.sdk.name || sdkInfo.name;
+  event.sdk.version = event.sdk.version || sdkInfo.version;
+  event.sdk.integrations = [...(event.sdk.integrations || []), ...(sdkInfo.integrations || [])];
+  event.sdk.packages = [...(event.sdk.packages || []), ...(sdkInfo.packages || [])];
+  return event;
+}
+
 /** Creates a SentryRequest from an event. */
 export function sessionToSentryRequest(session: Session, api: API): SentryRequest {
+  const sdkInfo = getSdkMetadataForEnvelopeHeader(api);
   const envelopeHeaders = JSON.stringify({
     sent_at: new Date().toISOString(),
+    ...(sdkInfo && { sdk: sdkInfo }),
   });
   const itemHeaders = JSON.stringify({
     type: 'session',
@@ -20,11 +51,17 @@ export function sessionToSentryRequest(session: Session, api: API): SentryReques
 
 /** Creates a SentryRequest from an event. */
 export function eventToSentryRequest(event: Event, api: API): SentryRequest {
-  const useEnvelope = event.type === 'transaction';
+  // since JS has no Object.prototype.pop()
+  const { __sentry_samplingMethod: samplingMethod, __sentry_sampleRate: sampleRate, ...otherTags } = event.tags || {};
+  event.tags = otherTags;
+
+  const sdkInfo = getSdkMetadataForEnvelopeHeader(api);
+  const eventType = event.type || 'event';
+  const useEnvelope = eventType === 'transaction';
 
   const req: SentryRequest = {
-    body: JSON.stringify(event),
-    type: event.type || 'event',
+    body: JSON.stringify(sdkInfo ? enhanceEventWithSdkInfo(event, api.metadata.sdk) : event),
+    type: eventType,
     url: useEnvelope ? api.getEnvelopeEndpointWithUrlEncodedAuth() : api.getStoreEndpointWithUrlEncodedAuth(),
   };
 
@@ -38,9 +75,15 @@ export function eventToSentryRequest(event: Event, api: API): SentryRequest {
     const envelopeHeaders = JSON.stringify({
       event_id: event.event_id,
       sent_at: new Date().toISOString(),
+      ...(sdkInfo && { sdk: sdkInfo }),
     });
     const itemHeaders = JSON.stringify({
       type: event.type,
+
+      // TODO: Right now, sampleRate may or may not be defined (it won't be in the cases of inheritance and
+      // explicitly-set sampling decisions). Are we good with that?
+      sample_rates: [{ id: samplingMethod, rate: sampleRate }],
+
       // The content-type is assumed to be 'application/json' and not part of
       // the current spec for transaction items, so we don't bloat the request
       // body with it.