+++ packages/tracing/src/browser/index.ts
@@ -1,1 +1,6 @@
 export { BrowserTracing } from './browsertracing';
+export {
+  registerRequestInstrumentation,
+  RequestInstrumentationOptions,
+  defaultRequestInstrumentationOptions,
+} from './request';