+++ packages/types/src/index.ts
@@ -2,6 +2,7 @@ export { Breadcrumb, BreadcrumbHint } from './breadcrumb';
 export { Client } from './client';
 export { Context, Contexts } from './context';
 export { Dsn, DsnComponents, DsnLike, DsnProtocol } from './dsn';
+export { DebugImage, DebugImageType, DebugMeta } from './debugMeta';
 export { ExtendedError } from './error';
 export { Event, EventHint } from './event';
 export { EventProcessor } from './eventprocessor';
@@ -11,7 +12,7 @@ export { Hub } from './hub';
 export { Integration, IntegrationClass } from './integration';
 export { LogLevel } from './loglevel';
 export { Mechanism } from './mechanism';
-export { ExtractedNodeRequestData, WorkerLocation } from './misc';
+export { ExtractedNodeRequestData, Primitive, WorkerLocation } from './misc';
 export { Options } from './options';
 export { Package } from './package';
 export { Request, SentryRequest, SentryRequestType } from './request';
@@ -19,6 +20,7 @@ export { Response } from './response';
 export { Runtime } from './runtime';
 export { CaptureContext, Scope, ScopeContext } from './scope';
 export { SdkInfo } from './sdkinfo';
+export { SdkMetadata } from './sdkmetadata';
 export { Session, SessionContext, SessionStatus } from './session';
 export { Severity } from './severity';
 export { Span, SpanContext } from './span';
@@ -32,6 +34,7 @@ export {
   TraceparentData,
   Transaction,
   TransactionContext,
+  TransactionSamplingMethod,
 } from './transaction';
 export { Thread } from './thread';
 export { Transport, TransportOptions, TransportClass } from './transport';