+++ packages/types/src/stackframe.ts
@@ -11,5 +11,7 @@ export interface StackFrame {
   pre_context?: string[];
   post_context?: string[];
   in_app?: boolean;
+  instruction_addr?: string;
+  addr_mode?: string;
   vars?: { [key: string]: any };
 }